Fixing problems created by moving PR to new repository
