+++ libs/core/algorithms/include/hpx/parallel/datapar/adjacent_find.hpp
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <hpx/config.hpp>
+#include <hpx/local/config.hpp>
 
 #if defined(HPX_HAVE_DATAPAR)
 #include <hpx/concepts/concepts.hpp>
